requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bootstrap4==0.1.0
boto3==1.24.26
botocore==1.27.26
class-registry==2.1.2
Django==4.0.5
django-bootstrap==0.2.4
django-bootstrap4==22.1
django-cors-headers==3.13.0
django-forms-bootstrap==3.1.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filters==1.3.2
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
regex==2022.6.2
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
uuid==1.30
whitenoise==6.2.0
